Refactored shifters into package, improved Double...
