+++ src/org/team5940/robot_core/modules/functional/shifters/ShifterModule.java
@@ -1,4 +1,4 @@
-package org.team5940.robot_core.modules.functional;
+package org.team5940.robot_core.modules.functional.shifters;
 
 import org.team5940.robot_core.modules.ownable.OwnableModule;
 /**
